+++ arcade/sprite_list/__init__.py
@@ -1,5 +1,6 @@
-from .sprite_list import SpriteList  # noqa
-from .spatial_hash import (  # noqa
+from .sprite_list import SpriteList
+from .spatial_hash import SpatialHash
+from .collision import (
     get_closest_sprite,
     check_for_collision,
     check_for_collision_with_list,
@@ -7,3 +8,15 @@ from .spatial_hash import (  # noqa
     get_sprites_at_point,
     get_sprites_at_exact_point,
 )
+
+
+__all__ = [
+    "SpriteList",
+    "SpatialHash",
+    "get_closest_sprite",
+    "check_for_collision",
+    "check_for_collision_with_list",
+    "check_for_collision_with_lists",
+    "get_sprites_at_point",
+    "get_sprites_at_exact_point",
+]
